requirements.txt
fastapi
uvicorn
django
httpx
requests
playwright
sqlalchemy
python-decouple
pymysql